pyproject.toml
[build-system]
requires = ['flit_core>=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'aiida-vibroscopy'
dynamic = ['description', 'version']
authors = [
    {name = 'Lorenzo Bastonero', email = 'bastonero.lorenzo@gmail.com'}
]
readme = 'README.md'
license = {file = 'LICENSE.txt'}
classifiers = [
    'Development Status :: 4 - Beta',
    'Framework :: AiiDA',
    'License :: OSI Approved :: MIT License',
    'Operating System :: POSIX :: Linux',
    'Operating System :: MacOS :: MacOS X',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
keywords = ['aiida', 'workflows']
requires-python = '>=3.8'
dependencies = [
    "aiida-core>=2.2.2",
    "aiida-quantumespresso>=4.3.0",
    "aiida-phonopy>=1.1.2",
    "phonopy>=2.16.0,<3.0.0",
]

[project.urls]
Source = "https://github.com/bastonero/aiida-vibroscopy"

[project.optional-dependencies]
pre-commit = [
    'pre-commit~=2.17',
    'pylint==2.13.7',
    'pylint-aiida~=0.1.1',
    'toml',
]
tests = [
    'pgtest~=1.3',
    'pytest~=6.0',
    'pytest-regressions~=2.3',
]
docs = [
    'myst-nb~=0.17',
    'jupytext>=1.11.2,<1.15.0',
    'sphinx-togglebutton',
    'sphinx~=5.2',
    'sphinx-copybutton~=0.5.2',
    'sphinx-book-theme~=1.0.1',
    'sphinx-design~=0.4.1',
    'sphinxcontrib-details-directive~=0.1.0',
    'sphinx-autoapi~=2.0.1'
]

[project.entry-points.'aiida.data']
"vibroscopy.fp" = "aiida_vibroscopy.data.vibro_fp:VibrationalFrozenPhononData"
"vibroscopy.vibrational" =  "aiida_vibroscopy.data.vibro_lr:VibrationalData"

[project.entry-points.'aiida.workflows']
"vibroscopy.dielectric" = "aiida_vibroscopy.workflows.dielectric.base:DielectricWorkChain"
"vibroscopy.dielectric.numerical_derivatives" = "aiida_vibroscopy.workflows.dielectric.numerical_derivatives:NumericalDerivativesWorkChain"
"vibroscopy.phonons.harmonic" = "aiida_vibroscopy.workflows.phonons.harmonic:HarmonicWorkChain"
"vibroscopy.phonons.phonon" = "aiida_vibroscopy.workflows.phonons.base:PhononWorkChain"
"vibroscopy.spectra.iraman" = "aiida_vibroscopy.workflows.spectra.iraman:IRamanSpectraWorkChain"
"vibroscopy.spectra.intensities_average" = "aiida_vibroscopy.workflows.spectra.intensities_average:IntensitiesAverageWorkChain"

[tool.flit.module]
name = 'aiida_vibroscopy'

[tool.flit.sdist]
exclude = [
    'docs/',
    'tests/',
    '.github/',
    '.gitignore',
    '.pre-commit-config.yaml',
]

[tool.flynt]
line-length = 120
fail-on-change = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.pydocstyle]
ignore = [
    'D104',
    'D203',
    'D213'
]

[tool.pylint.basic]
good-names = [
    "_",
    "x",
    "y",
    "z",
    "i",
    "j",
    "k",
    "l",
    "dph0_",
    "nlo_",
]

[tool.pylint.master]
load-plugins = ['pylint_aiida']
ignore-paths= [
    'tests/*',
    'docs/*',
]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    'protected-access',
    'duplicate-code',
    'import-outside-toplevel',
    'inconsistent-return-statements',
    'invalid-name',
    'no-member',
    'too-many-arguments',
    'too-many-branches',
    'too-many-locals',
    'too-many-statements',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 120
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true
